+++ ServerApp/Interfaces/IService.cs
@@ -5,9 +5,9 @@ namespace ServerApp.Interfaces
     public interface IService<T>
     {
         Task<List<T>> Get();
-        Task<T?> GetById(int id);
+        Task<T?> GetById(long id);
         Task<bool> Add(T entity);
-        Task<bool> Update(int id, T entity);
-        Task<bool> DeleteById(int id);
+        Task<bool> Update(long id, T entity);
+        Task<bool> DeleteById(long id);
     }
 }
